Scope notification title refreshes to changed workspaces
