+++ Sources/Update/NotificationsPopoverMetrics.swift
@@ -1,0 +1,10 @@
+import CoreGraphics
+
+enum NotificationsPopoverMetrics {
+    static let defaultWidth: CGFloat = 560
+    static let defaultHeight: CGFloat = 760
+    static let minWidth: CGFloat = 420
+    static let minHeight: CGFloat = 320
+    static let maxWidth: CGFloat = 1000
+    static let maxHeight: CGFloat = 1200
+}
