+++ Sources/Update/UpdateTitlebarAccessory.swift
@@ -2130,15 +2130,6 @@ final class TitlebarControlsAccessoryViewController: NSTitlebarAccessoryViewCont
     }
 }
 
-private enum NotificationsPopoverMetrics {
-    static let defaultWidth: CGFloat = 560
-    static let defaultHeight: CGFloat = 760
-    static let minWidth: CGFloat = 420
-    static let minHeight: CGFloat = 320
-    static let maxWidth: CGFloat = 1000
-    static let maxHeight: CGFloat = 1200
-}
-
 private struct NotificationsPopoverView: View {
     @ObservedObject var notificationStore: TerminalNotificationStore
     @ObservedObject private var keyboardShortcutSettingsObserver = KeyboardShortcutSettingsObserver.shared
@@ -2177,6 +2168,9 @@ private struct NotificationsPopoverView: View {
             refreshWorkspaceTitles(ifRelevantTo: notification)
         }
         .onReceive(NotificationCenter.default.publisher(for: .workspaceGroupNameDidChange)) { notification in
+            refreshWorkspaceTitles(ifRelevantTo: notification)
+        }
+        .onReceive(NotificationCenter.default.publisher(for: .workspaceOrderDidChange)) { notification in
             refreshWorkspaceTitles(ifRelevantTo: notification)
         }
     }
@@ -2392,27 +2386,49 @@ private struct NotificationsPopoverView: View {
 
     private func currentWorkspaceTitles() -> [UUID: String] {
         let notificationWorkspaceIds = Set(notificationStore.notifications.map(\.tabId))
-        return (AppDelegate.shared?.tabTitlesByTabId() ?? [:]).filter {
-            notificationWorkspaceIds.contains($0.key)
-        }
+        return AppDelegate.shared?.tabTitlesByTabId(for: notificationWorkspaceIds) ?? [:]
     }
 
     private func refreshWorkspaceTitles(ifRelevantTo notification: Notification? = nil) {
         let notificationWorkspaceIds = Set(notificationStore.notifications.map(\.tabId))
-        if let notification {
-            switch notification.name {
-            case .workspaceTitleDidChange:
-                guard let workspaceId = notification.userInfo?[GhosttyNotificationKey.tabId] as? UUID,
-                      notificationWorkspaceIds.contains(workspaceId) else { return }
-            case .workspaceGroupNameDidChange:
-                guard let manager = notification.object as? TabManager,
-                      manager.tabs.contains(where: { notificationWorkspaceIds.contains($0.id) }) else { return }
-            default:
-                return
-            }
+        guard let notification else {
+            let nextTitles = currentWorkspaceTitles()
+            guard nextTitles != workspaceTitles else { return }
+            workspaceTitles = nextTitles
+            return
         }
 
-        let nextTitles = currentWorkspaceTitles()
+        guard let manager = notification.object as? TabManager else { return }
+        let changedWorkspaceIds: Set<UUID>
+        let changedTitles: [UUID: String]
+        switch notification.name {
+        case .workspaceTitleDidChange:
+            guard let workspaceId = notification.userInfo?[GhosttyNotificationKey.tabId] as? UUID else { return }
+            changedWorkspaceIds = [workspaceId]
+            changedTitles = manager.resolvedWorkspaceDisplayTitle(forWorkspaceId: workspaceId)
+                .map { [workspaceId: $0] } ?? [:]
+        case .workspaceGroupNameDidChange:
+            changedTitles = manager.resolvedWorkspaceDisplayTitles(for: notificationWorkspaceIds)
+            changedWorkspaceIds = Set(changedTitles.keys)
+        case .workspaceOrderDidChange:
+            changedWorkspaceIds = Set(
+                notification.userInfo?[WorkspaceOrderChangeNotificationKey.movedWorkspaceIds] as? [UUID] ?? []
+            )
+            changedTitles = manager.resolvedWorkspaceDisplayTitles(for: changedWorkspaceIds)
+        default:
+            return
+        }
+
+        let relevantIds = changedWorkspaceIds.intersection(notificationWorkspaceIds)
+        guard !relevantIds.isEmpty else { return }
+        var nextTitles = workspaceTitles
+        for workspaceId in relevantIds {
+            if let title = changedTitles[workspaceId] {
+                nextTitles[workspaceId] = title
+            } else {
+                nextTitles.removeValue(forKey: workspaceId)
+            }
+        }
         guard nextTitles != workspaceTitles else { return }
         workspaceTitles = nextTitles
     }
@@ -2462,89 +2478,6 @@ private struct NotificationsPopoverView: View {
         DispatchQueue.main.async {
             _ = AppDelegate.shared?.openTerminalNotification(notification)
             onDismiss()
-        }
-    }
-}
-
-struct HoverTrackingRepresentable: NSViewRepresentable {
-    let onChange: (Bool) -> Void
-
-    func makeNSView(context: Context) -> HoverTrackingNSView {
-        HoverTrackingNSView(onChange: onChange)
-    }
-
-    func updateNSView(_ nsView: HoverTrackingNSView, context: Context) {
-        nsView.onChange = onChange
-    }
-}
-
-final class HoverTrackingNSView: NSView {
-    var onChange: (Bool) -> Void
-    private var trackingArea: NSTrackingArea?
-    private var isInside: Bool = false
-
-    init(onChange: @escaping (Bool) -> Void) {
-        self.onChange = onChange
-        super.init(frame: .zero)
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-
-    // Pass clicks through to the SwiftUI parent (which owns the tap gesture and accessibility
-    // action). Tracking areas keep working because they're driven by window mouse-tracking,
-    // not by hitTest.
-    override func hitTest(_ point: NSPoint) -> NSView? { nil }
-
-    override func updateTrackingAreas() {
-        super.updateTrackingAreas()
-        if let trackingArea {
-            removeTrackingArea(trackingArea)
-        }
-        let area = NSTrackingArea(
-            rect: .zero,
-            options: [.mouseEnteredAndExited, .mouseMoved, .activeAlways, .inVisibleRect],
-            owner: self,
-            userInfo: nil
-        )
-        addTrackingArea(area)
-        trackingArea = area
-
-        // Sync current pointer state in case the pointer is already inside when the tracking
-        // area is (re)installed — happens on first popover open or after layout changes.
-        // updateTrackingAreas runs on the main thread, so dispatch synchronously; deferring
-        // creates a race where mouseExited can fire before the queued sync-onChange(true) runs,
-        // leaving the row stuck in the hovered state.
-        if let window, window.isVisible {
-            let mouseInWindow = window.mouseLocationOutsideOfEventStream
-            let mouseInView = convert(mouseInWindow, from: nil)
-            let nowInside = bounds.contains(mouseInView)
-            if nowInside != isInside {
-                isInside = nowInside
-                onChange(nowInside)
-            }
-        }
-    }
-
-    override func mouseEntered(with event: NSEvent) {
-        if !isInside {
-            isInside = true
-            onChange(true)
-        }
-    }
-
-    override func mouseExited(with event: NSEvent) {
-        if isInside {
-            isInside = false
-            onChange(false)
-        }
-    }
-
-    override func viewDidMoveToWindow() {
-        super.viewDidMoveToWindow()
-        if window == nil, isInside {
-            isInside = false
-            onChange(false)
         }
     }
 }
